Deleted unnecesary files and made app to start
Deleted unnecesary / conflicting files and made app to start to log in screen.
