+++ IT-M/Mobile/App.js
@@ -48,6 +48,6 @@ export default function App() {
     }
   };
 
-  return <AppNavigator />;
+  return <AppNavigator onLogin={handleLogin} />;
 }
