+++ IT-M/Mobile/navigation/AppNavigator.js
@@ -1,0 +1,47 @@
+import React from "react";
+import { View, Text, StyleSheet } from "react-native";
+import { NavigationContainer } from "@react-navigation/native";
+import { createStackNavigator } from "@react-navigation/stack";
+import Login from "../pages/Login";
+
+const Stack = createStackNavigator();
+
+function WelcomeScreen() {
+  return (
+    <View style={styles.container}>
+      <Text style={styles.title}>Welcome</Text>
+      <Text style={styles.subtitle}>You are logged in.</Text>
+    </View>
+  );
+}
+
+export default function AppNavigator({ onLogin }) {
+  return (
+    <NavigationContainer>
+      <Stack.Navigator initialRouteName="Login">
+        <Stack.Screen name="Login">
+          {(props) => <Login {...props} onLogin={onLogin} />}
+        </Stack.Screen>
+        <Stack.Screen name="Welcome" component={WelcomeScreen} />
+      </Stack.Navigator>
+    </NavigationContainer>
+  );
+}
+
+const styles = StyleSheet.create({
+  container: {
+    flex: 1,
+    alignItems: "center",
+    justifyContent: "center",
+    padding: 24,
+  },
+  title: {
+    fontSize: 24,
+    fontWeight: "600",
+    marginBottom: 8,
+  },
+  subtitle: {
+    fontSize: 16,
+    color: "#475569",
+  },
+});
