+++ IT-M/Mobile/package.json
@@ -17,6 +17,7 @@
     "axios": "^1.13.5",
     "react": "19.2.0",
     "react-native": "0.83.1",
+    "react-native-gesture-handler": "^2.20.2",
     "react-native-safe-area-context": "^5.6.2",
     "react-native-screens": "^4.23.0"
   },
