--- Mobile/package.json
@@ -1,29 +1,0 @@
-{
-  "name": "mobile",
-  "version": "1.0.0",
-  "main": "index.ts",
-  "scripts": {
-    "start": "expo start",
-    "android": "expo start --android",
-    "ios": "expo start --ios"
-  },
-  "dependencies": {
-    "@react-navigation/native": "^6.1.18",
-    "@react-navigation/native-stack": "^6.9.26",
-    "@tanstack/react-query": "^5.90.20",
-    "axios": "^1.13.4",
-    "expo": "~50.0.17",
-    "expo-secure-store": "~12.8.1",
-    "expo-status-bar": "~1.11.1",
-    "react": "18.2.0",
-    "react-native": "0.73.6",
-    "react-native-safe-area-context": "4.8.2",
-    "react-native-screens": "~3.29.0"
-  },
-  "devDependencies": {
-    "@types/react": "~18.2.21",
-    "@types/react-native": "~0.73.0",
-    "typescript": "~5.2.2"
-  },
-  "private": true
-}
