--- Mobile/src/api/client.ts
@@ -1,14 +1,0 @@
-import axios from "axios";
-import * as SecureStore from "expo-secure-store";
-
-export const api = axios.create({
-  baseURL: "https://api.luckybunny.eu/api",
-});
-
-api.interceptors.request.use(async (config) => {
-  const token = await SecureStore.getItemAsync("token");
-  if (token) {
-    config.headers.Authorization = `Bearer ${token}`;
-  }
-  return config;
-});
